Moved NOS3 integration tests to standalone files. Disabled running on examples in docs.
