+++ Cargo.toml
@@ -47,6 +47,7 @@ members = [
 "test/integration/linux/mai400",
 "test/integration/large_download",
 "test/integration/large_upload",
+"test/integration/nosengine-rust",
 "utils"
 ]
