+++ test/integration/nosengine-rust/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "nosengine-integration-tests"
+version = "0.1.0"
+authors = ["Ryan Plauche <ryan@kubos.co>"]
+edition = "2018"
+
+[dependencies]
+nosengine-rust = { path = "../../../apis/nosengine-rust" }
+utils = { path = "../../../utils" }
+
