+++ test/integration/nosengine-rust/sim_config4.json
@@ -1,0 +1,16 @@
+{
+    "plugins": [
+        "uart"
+    ],
+
+    "server_uris": [
+        {
+            "name": "fsw",
+            "server_uri": "tcp://127.0.0.4:12000"
+        },
+        {
+            "name": "nos3",
+            "server_uri": "tcp://127.0.0.4:12001"
+        }
+    ]
+}
